Separate malaphor-generating function into separate file to app.py
