+++ templates/index.html
@@ -7,6 +7,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link rel="stylesheet" href="{{ url_for('static', filename='styles.css') }}">
     <title>Malaphor Generator</title>
+    <link rel="icon" type="image/x-icon" href="{{ url_for('static', filename='img/favicon.png') }}">
 </head>
 
 <body>
@@ -17,7 +18,6 @@
     </div>
     <div class="malaphor-container">
         <h1 id="page-start">Malaphor Generator</h1>
-
 <div id="generated-malaphors">
   <em><p class="context_text">The idioms being merged are:</p></em>
     {% if currentIdiom %}
